upgrade gen instances

+++ src/main_gen_inst.cpp
@@ -6,7 +6,7 @@ void gen_graphs_for_stats() {
     std::ofstream csv("../data.csv"); // on va écrire les résultats obtenus dedans 
 
     for(int n = 100; n < 2000; n+=100) {
-        for(int p = 0.05; p < 1.0; p += 0.05) {
+        for(float p = 0.05; p < 1.0; p += 0.05) {
             for(int test = 0; test < 10; ++test) {
 
                 std::cout << "n : " << n << " / ";
